requirements.txt
Flask==3.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
